fix string literals (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@953 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/dialog/d_clnd.cpp
@@ -108,7 +108,7 @@ LRESULT CALLBACK ClndDialogProc(HWND hWnd, UINT msg, WPARAM wp, LPARAM lp) {
 		case WM_COMMAND:
 			switch(LOWORD(wp)) {
 				case IDOK:
-					b = GetDlgItemCheck(hWnd, IDC_CLNDREAL);
+					b = (UINT8)GetDlgItemCheck(hWnd, IDC_CLNDREAL);
 					if (np2cfg.calendar != b) {
 						np2cfg.calendar = b;
 						sysmng_update(SYS_UPDATECFG);
